UPD: move load image function to detector_utils.py
